refactored layouts to fit gatsby convention
