+++ src/layouts/index.css
@@ -21,3 +21,18 @@ section p.cta {
 .clear-fix {
   clear: both;
 }
+
+.layout {
+  background-color: #ffffff;
+  width: 80%;
+  margin: 0 auto 20px;
+}
+
+.layout hr {
+  width: 15%;
+  margin: 5px auto;
+}
+
+.layout .outlet {
+  margin: 0 0 20px;
+}
