--- src/layouts/page-layout.css
@@ -1,14 +1,0 @@
-.layout {
-  background-color: #ffffff;
-  width: 80%;
-  margin: 0 auto 20px;
-}
-
-.layout hr {
-  width: 15%;
-  margin: 5px auto;
-}
-
-.layout .outlet {
-  margin: 0 0 20px;
-}
